revert: "refactor!: Merge branch 'master' of https://github.com/tyabus/Kisak-Strike."
This reverts commit e619ba8c4ee0f66c38894d7a48a3d8be74b8382a, reversing
changes made to 5cd21d5f708f4dcee49851a3cfa917009ef192c3.

+++ vgui2/vgui_controls/FileOpenStateMachine.cpp
@@ -9,8 +9,10 @@
 #include "tier1/keyvalues.h"
 #include "vgui_controls/FileOpenDialog.h"
 #include "vgui_controls/MessageBox.h"
+#include "vgui_controls/perforcefilelistframe.h"
 #include "vgui_controls/savedocumentquery.h"
 #include "filesystem.h"
+#include "p4lib/ip4.h"
 #include "tier2/tier2.h"
 #include "tier0/icommandline.h"
 
@@ -195,6 +197,12 @@ void FileOpenStateMachine::WriteFile()
 	}
 
 	m_bWroteFile = true;
+	if ( m_bShowPerforceDialogs )
+	{
+		m_CurrentState = STATE_SHOWING_PERFORCE_ADD_DIALOG;
+		ShowPerforceQuery( GetParent(), m_FileName, this, NULL, PERFORCE_ACTION_FILE_ADD );
+		return;
+	}
 
 	if ( !m_bIsOpeningFile )
 	{
@@ -238,6 +246,15 @@ void FileOpenStateMachine::MakeFileWriteableDialog( )
 		return;
 	}
 
+	// If it's in perforce, and not checked out, then we must abort.
+	bool bIsInPerforce = p4->IsFileInPerforce( m_FileName );
+	bool bIsOpened = ( p4->GetFileState( m_FileName ) != P4FILE_UNOPENED );
+	if ( bIsInPerforce && !bIsOpened )
+	{
+		SetCompletionState( FILE_NOT_CHECKED_OUT );
+		return;
+	}
+
 	m_CurrentState = STATE_SHOWING_MAKE_FILE_WRITEABLE_DIALOG;
 
 	char pBuf[1024];
@@ -272,6 +289,18 @@ void FileOpenStateMachine::OnPerforceQueryCompleted( KeyValues *pKeyValues )
 		return;
 	}
 
+	if ( m_CurrentState == STATE_SHOWING_PERFORCE_ADD_DIALOG )
+	{
+		if ( !m_bIsOpeningFile )
+		{
+			SetCompletionState( SUCCESSFUL );
+			return;
+		}
+
+		OpenFileDialog();
+		return;
+	}
+
 	Assert(0);
 }
 
@@ -281,6 +310,13 @@ void FileOpenStateMachine::OnPerforceQueryCompleted( KeyValues *pKeyValues )
 //-----------------------------------------------------------------------------
 void FileOpenStateMachine::CheckOutDialog( )
 {
+	if ( m_bShowPerforceDialogs )
+	{
+		m_CurrentState = STATE_SHOWING_CHECK_OUT_DIALOG;
+		ShowPerforceQuery( GetParent(), m_FileName, this, NULL, PERFORCE_ACTION_FILE_EDIT );
+		return;
+	}
+
 	WriteFile();
 }
 
@@ -345,6 +381,13 @@ void FileOpenStateMachine::SaveFile( KeyValues *pContextKeyValues, const char *p
 	m_OpenFileType = NULL;
 	m_OpenFileName = NULL;
 
+	// Clear the P4 dialog flag for SDK users and licensees without Perforce
+	if ( g_pFullFileSystem->IsSteam() || CommandLine()->FindParm( "-nop4" ) )
+	{
+		nFlags &= ~FOSM_SHOW_PERFORCE_DIALOGS;
+	}
+
+	m_bShowPerforceDialogs = ( nFlags & FOSM_SHOW_PERFORCE_DIALOGS ) != 0;
 	m_bShowSaveQuery = ( nFlags & FOSM_SHOW_SAVE_QUERY ) != 0;
 	m_bIsOpeningFile = false;
 	m_bWroteFile = false;
@@ -409,6 +452,7 @@ void FileOpenStateMachine::OpenFile( const char *pOpenFileType, KeyValues *pCont
 	m_SaveFileType = pSaveFileType;
 	m_OpenFileType = pOpenFileType;
 	m_OpenFileName = NULL;
+	m_bShowPerforceDialogs = ( nFlags & FOSM_SHOW_PERFORCE_DIALOGS ) != 0;
 	m_bShowSaveQuery = ( nFlags & FOSM_SHOW_SAVE_QUERY ) != 0;
 	m_bIsOpeningFile = true;
 	m_bWroteFile = false;
@@ -434,6 +478,7 @@ void FileOpenStateMachine::OpenFile( const char *pOpenFileName, const char *pOpe
 	m_FileName = pSaveFileName;
 	m_SaveFileType = pSaveFileType;
 	m_OpenFileType = pOpenFileType;
+	m_bShowPerforceDialogs = ( nFlags & FOSM_SHOW_PERFORCE_DIALOGS ) != 0;
 	m_bShowSaveQuery = ( nFlags & FOSM_SHOW_SAVE_QUERY ) != 0;
 	m_bIsOpeningFile = true;
 	m_bWroteFile = false;
